% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro5.R
\name{mol_table_construction}
\alias{mol_table_construction}
\title{Lipinskis Rule of 5 Function}
\usage{
mol_table_construction(compound_dataset)
}
\arguments{
\item{compound_dataset}{The input compound dataset}
}
\value{
New filtered compound dataset based on Rule of 5.
}
\description{
Lipinskis Rule of 5 Function
}
\examples{
mol_table_construction(compound_dataset)
\dontrun{
mol_table_construction(compound_dataset)
}
}
