% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro5.R
\name{radar_chart}
\alias{radar_chart}
\title{Radar Chart Summarising Compound Dataset}
\usage{
radar_chart(mols_filtered)
}
\arguments{
\item{mols_filtered}{}
}
\value{
A radar plot summarising Compound Dataset
}
\description{
Radar Chart Summarising Compound Dataset
}
\examples{
radar_chart(mols_filtered)
\dontrun{
radar_chart(mols_filtered)
}
}
