% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro5.R
\name{ro5filter}
\alias{ro5filter}
\title{Lipinskis Rule of 5 Filter}
\usage{
ro5filter(compound_dataset)
}
\arguments{
\item{compound_dataset}{The input compound dataset}
}
\value{
New filtered compound dataset based on Lipinskis Rule of 5.
}
\description{
Lipinskis Rule of 5 Filter
}
\examples{
ro5filter(compound_dataset)
\dontrun{
ro5filter(compound_dataset)
}
}
