% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro5.R
\name{pca}
\alias{pca}
\title{Compound Dataset PCA}
\usage{
pca(filtered_dataset)
}
\arguments{
\item{filtered_dataset}{Filtered compound dataset based on Rule of 5.}
}
\value{
PCA plot of filtered compound dataset.
}
\description{
Compound Dataset PCA
}
\examples{
pca(filtered_dataset)
\dontrun{
pca(filtered_dataset)
}
}
